requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
PyHamcrest==2.0.4
pytest==7.4.0
requests==2.31.0
urllib3==2.0.4
